requirements.txt
alembic==1.8.0
anyio==3.6.1
asgiref==3.5.2
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
et-xmlfile==1.1.0
fastapi==0.78.0
greenlet==1.1.2
grep==0.3.2
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
Naked==0.1.31
numpy==1.22.4
openpyxl==3.0.10
orjson==3.7.2
pandas==1.4.2
passlib==1.7.4
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.28.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
status==0.2.5
typing_extensions==4.2.0
ujson==5.3.0
urllib3==1.26.9
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
